requirements.txt
alembic==1.7.7
anyio==3.5.0
asgiref==3.5.0
asyncclick==8.0.3.2
asyncpg==0.25.0
bcrypt==3.2.0
black==22.3.0
boto3==1.21.32
botocore==1.24.32
cffi==1.15.0
click==8.1.0
databases==0.5.5
fastapi==0.75.0
greenlet==1.1.2
h11==0.13.0
httptools==0.4.0
idna==3.3
jmespath==1.0.0
Mako==1.2.0
MarkupSafe==2.1.1
mypy-extensions==0.4.3
passlib==1.7.4
pathspec==0.9.0
platformdirs==2.5.1
psycopg2-binary==2.9.3
pycparser==2.21
pydantic==1.9.0
PyJWT==2.3.0
python-dateutil==2.8.2
python-decouple==3.6
python-dotenv==0.20.0
PyYAML==6.0
s3transfer==0.5.2
six==1.16.0
sniffio==1.2.0
SQLAlchemy==1.4.32
starlette==0.17.1
tomli==2.0.1
typing_extensions==4.1.1
urllib3==1.26.9
uvicorn==0.17.6
uvloop==0.16.0
watchgod==0.8.1
websockets==10.2
